+++ R/top_by_go.R
@@ -92,12 +92,12 @@ setMethod("top_by_go",
               print_msg(paste0("Processing cluster ", i), msg_type = "INFO")
               
               go_list <- biomaRt::getBM(attributes=c("go_id",
-                                            "hgnc_symbol"),
-                               filters = "hgnc_symbol",
+                                                     gene_id),
+                               filters = gene_id,
                                values = object@gene_clusters[[i]],
                                mart=mart)
-              
-              to_keep <- go_list$hgnc_symbol[go_list$go_id %in% go_id]
+
+              to_keep <- go_list[,gene_id][go_list$go_id %in% go_id]
               top_gn[[i]] <- sort(to_keep)
             }